Update cd math master
Note: It is very important that you perform the git checkout update-0 command in order to follow along with the entire lesson. As stated earlier, this command resets the code in the repo to a point before applying any of the changes we are going to make in this lesson.